use ADT for strings

+++ src/ast/rewriter/seq_rewriter.h
@@ -58,7 +58,7 @@ class seq_rewriter {
                         expr_ref_vector& lhs, expr_ref_vector& rhs, bool& is_sat);
     bool length_constrained(unsigned n, expr* const* l, unsigned m, expr* const* r, 
                         expr_ref_vector& lhs, expr_ref_vector& rhs, bool& is_sat);
-    bool min_length(unsigned n, expr* const* es, size_t& len);
+    bool min_length(unsigned n, expr* const* es, unsigned& len);
     expr* concat_non_empty(unsigned n, expr* const* es);
 
 public:
